finish the changes I was aiming for:
 - event changes via epoll_ctl() are now delayed until just before
epoll_wait() so that redundant changes don't result in multiple calls
and so that calls that set both read and write events are combined into
one.

 - queue fd changes so that if we have a large number of file handles
with activity only on a few we avoid a linear search of the fd array.

This version has one bug: the fd is removed from the fds array when
want_events is this, but this causes a warning if pause is used to
pause reads, and ignore is used to ignore writes, deleting the handle
before ignoring writes